Restructured dart files hierarchy
